+++ Vector2D.h
@@ -25,7 +25,7 @@ public:
 	Vector2D& operator*=(const Vector2D& vec);
 	Vector2D& operator/=(const Vector2D& vec);
 
-	Vector2D& operator*(const int& i);
+	Vector2D& operator*(const float& i);
 	Vector2D& Zero();
 
 	friend std::ostream& operator<<(std::ostream& stream, const Vector2D& vec);
